Emit new child class in @Module decorator
Instead of reusing the prototype of the constructor to be decorated
directly, we create a child class and thus keep the whole original class
intact.

Previously any metadata on the undecorated module class was lost. Now it
lives on in the prototype chain.
Cf. https://github.com/rbuckton/reflect-metadata

Fixes #9